add HannanBot (#6)
* Created bot

* Decided strategy

Decided on my strategy for the battlebot arena. Will go to closest corner and start shooting. Will implement tmrw or tongiht depending how busy I am

* updating

* added corners

figured out how to make the corners recognizable. working on getting the bot there at the start

* Detects closest corner now

Working on getting it into newRound when it starts looking for the closest corner and getting the actual movement in getMove but for now the bot just detects the closest corner

* Shooting logic

Cleaned up the code a bit and now I'm trying to get the shooting logic to work

* Bot should be ready for round 1

My bot's code seems to me like it's ready fro round 1, but for some reason battlebotarena isn't working from my laptop so I can't test it to be sure

* Done round1 everythign

completed the bot for round1